MFC r312418,r312422:
r312418:

Conditionalize hyperv support in gettimeofday(2) based on MK_HYPERV

The effect at runtime is negligible as the hyperv timer isn't available
except when hyperv is loaded.

This is a prerequisite for conditionalizing the header build/install out
of the build

r312422:

Only conditionally add in hyperv support if we're building amd64

This unbreaks the build because the assembly is written for x64.

Pointyhat to:	ngie


git-svn-id: https://svn.freebsd.org/base/stable/11@313539 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f