MFC r311950 (by bz):
Get rid of a compiler warning which I saw too often.
Include netinet/in.h before ip_compat.t which will then check if
IPPROTO_IPIP is defined or not.  Doing it the other way round,
ip_compat.h would not find it defined and netinet/in.h then
redefine it.


git-svn-id: https://svn.freebsd.org/base/stable/10@313648 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f