Move CTIO waitq from per-LUN to per-channel.
All resources lack of which may put CTIO into the queue are either
per-channel or potentially per-queue, but none of them are per-LUN.

This is a first step to fix live LUN disabling.  Before this change
any CTIOs held in a queue in time of disabling were just leaked.

MFC after:	2 weeks


git-svn-id: https://svn.freebsd.org/base/head@313936 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f